backend/requirements.txt
flask
flask-cors
llama-index
llama-index-vector-stores-qdrant
qdrant-client
python-dotenv
gunicorn
openai